requirements.txt
pandas
numpy
scikit-learn
transformers
torch
python-telegram-bot
python-dotenv

